feat: renamed folder and added logic for /anime
Folder was renamed to `anilist-api` for clarity, because even though it's not @/app/api nor @/pages/api, this way it's easier to identify.

Separated in server and client the exports, so there are not issues.

If, for example, @/anilist-api/anime exported everything, and we made use of something in a RSC, that makes the rest not usable for client side components, even if not using the same function.

Documented in @/anilist-api/docs.md.